feature: new dashboard routes
